requirements.txt
django==4.2.8
psycopg2-binary==2.9.9
environs==10.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
flake8==6.1.0
flake8-builtins==2.2.0
flake8-django==1.4
flake8-import-order==0.18.2
flake8-print==5.0.0
drf-yasg==1.21.7
